REFACTOR - Setup proyecto
Se refactoriza el setup para cargar las dependencias de cada módulo por inyección de dependencias y que quede más limpio, además de agrupar las dependencias por tipo y módulo.

Se modifica el setup de token para que dure solo 60 minutos su validez y además se incorpora un lifetime validatorque comprueba valores como not before y expires con la fecha actual.
