+++ elearning/appsettings.Development.json
@@ -8,7 +8,7 @@
   },
   "AppSettings": {
     "Secret": "0vTXplbiIiyDh2eEKId0WYYoofzpyoRMWeE0Wc37NN6q8L6uj3bjD0LLEkbqGNtJWdOwdfqIAcV7UTn65Haewyju1WUirO8X0",
-    "TokenLifetimeInMinutes": 10080
+    "TokenLifetimeInMinutes": 60
   },
   "ConnectionStrings": {
     "DefaultConnection": "server=localhost;database=elearning;user=root;password=CraftCode2023"
