+++ elearning/src/Shared/Infrastructure/Framework/Startup/Service/AuthenticationService.cs
@@ -1,0 +1,56 @@
+﻿using System;
+using System.Text;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.IdentityModel.Tokens;
+
+namespace elearning.src.Shared.Infrastructure.Framework.Startup.Service
+{
+    public class AuthenticationService : IService
+    {
+        private readonly IConfiguration configuration;
+
+        public AuthenticationService(IConfiguration configuration)
+        {
+            this.configuration = configuration;
+        }
+
+        public void Load(IServiceCollection services)
+        {
+            var appSettingsSection = configuration.GetSection("AppSettings");
+            var secret = configuration.GetValue<string>("AppSettings:Secret");
+            var key = Encoding.ASCII.GetBytes(secret);
+
+            services.AddAuthentication(x =>
+                {
+                    x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                    x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+                })
+                .AddJwtBearer(x =>
+                {
+                    x.RequireHttpsMetadata = false;
+                    x.SaveToken = true;
+                    x.TokenValidationParameters = new TokenValidationParameters
+                    {
+                        ValidateIssuerSigningKey = true,
+                        IssuerSigningKey = new SymmetricSecurityKey(key),
+                        ValidateIssuer = false,
+                        ValidateAudience = false,
+
+                        ValidateLifetime = true,
+                        LifetimeValidator = (
+                            DateTime? notBefore,
+                            DateTime? expires,
+                            SecurityToken securityToken,
+                            TokenValidationParameters validationParameters
+                        ) => {
+                            return notBefore <= DateTime.UtcNow && expires >= DateTime.UtcNow;
+                        }
+                    };
+                });
+
+            services.AddOptions();
+        }
+    }
+}
