+++ elearning/src/Shared/Infrastructure/Framework/Startup/Service/IService.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace elearning.src.Shared.Infrastructure.Framework.Startup.Service
+{
+    public interface IService
+    {
+        void Load(IServiceCollection services);
+    }
+}
