+++ elearning/src/Shared/Infrastructure/Framework/Startup/Service/SwaggerService.cs
@@ -1,0 +1,52 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.ApiExplorer;
+using Microsoft.AspNetCore.Mvc.Versioning;
+using Microsoft.Extensions.DependencyInjection;
+using Swashbuckle.AspNetCore.Filters;
+using Swashbuckle.AspNetCore.Swagger;
+
+namespace elearning.src.Shared.Infrastructure.Framework.Startup.Service
+{
+    public class SwaggerService : IService
+    {
+        public void Load(IServiceCollection services)
+        {
+            services.AddMvcCore().AddVersionedApiExplorer(o => o.GroupNameFormat = "'v'VVV");
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddApiVersioning(config =>
+            {
+                config.ReportApiVersions = true;
+                config.AssumeDefaultVersionWhenUnspecified = true;
+                config.DefaultApiVersion = new ApiVersion(1, 0);
+                config.ApiVersionReader = new HeaderApiVersionReader("api-version");
+            });
+
+            services.AddSwaggerGen(
+                options => {
+                    options.AddSecurityDefinition("oauth2", new ApiKeyScheme
+                    {
+                        Description = "Standard Authorization header using the Bearer scheme. Example: \"bearer {token}\"",
+                        In = "header",
+                        Name = "Authorization",
+                        Type = "apiKey"
+                    });
+
+                    options.OperationFilter<SecurityRequirementsOperationFilter>();
+
+                    var provider = services.BuildServiceProvider().GetRequiredService<IApiVersionDescriptionProvider>();
+
+                    foreach (var description in provider.ApiVersionDescriptions)
+                    {
+                        options.SwaggerDoc(
+                            description.GroupName,
+                            new Info()
+                            {
+                                Title = $"Sample API {description.ApiVersion}",
+                                Version = description.ApiVersion.ToString()
+                            });
+                    }
+                }
+            );
+        }
+    }
+}
