+++ elearning/src/Shared/Infrastructure/Framework/Startup/Subscriber/ISubscriber.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Http;
+
+namespace elearning.src.Shared.Infrastructure.Framework.Startup.Subscriber
+{
+    public interface ISubscriber
+    {
+        void Setup(HttpContext context);
+    }
+}
